requirements.txt
# ScrapeClaw - Enterprise Web Scraper
# Core (already installed in this environment)
httpx>=0.28.0
beautifulsoup4>=4.14.0
lxml>=6.0.0
openpyxl>=3.1.0
rich>=14.0.0
typer>=0.24.0

# Must install
anthropic>=0.40.0
tenacity>=9.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
fake-useragent>=1.5.0

# Optional: JS rendering
# playwright>=1.40.0
